# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Муршрут — персональная афиша
  summary: API персональных рекомендаций и готовых маршрутов на вечер
  description: Backend мобильного MAX mini-app «Муршрут». Для защищённых методов сначала получите JWT
    через раздел auth, затем нажмите Authorize.
  version: 1.0.0
paths:
...
  /api/v1/users/me:
    get:
      tags:
      - profile
      summary: Получить текущего пользователя
      operationId: me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
  /api/v1/users/me/preferences:
    get:
      tags:
      - profile
      summary: Получить предпочтения
      operationId: preferences_api_v1_users_me_preferences_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/PreferencesOut'
                - type: 'null'
                title: Response Preferences Api V1 Users Me Preferences Get
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
    put:
      tags:
      - profile
      summary: Сохранить предпочтения
      description: Завершает анкету и обновляет город текущего пользователя.
      operationId: save_preferences_api_v1_users_me_preferences_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Save Preferences Api V1 Users Me Preferences Put
        '502':
          description: Внешний поставщик событий временно недоступен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
  /api/v1/recommendations:
    get:
      tags:
      - events
      summary: Получить персональные рекомендации
      description: Не возвращает события, на которые пользователь уже отреагировал.
      operationId: recommendations_api_v1_recommendations_get
      security:
      - BearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Максимальное число рекомендаций.
          default: 20
          title: Limit
        description: Максимальное число рекомендаций.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventOut'
                title: Response Recommendations Api V1 Recommendations Get
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Операция недоступна до завершения анкеты.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Recommendations Api V1 Recommendations Get
        '502':
          description: Внешний поставщик событий временно недоступен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
  /api/v1/events/catalog:
    get:
      tags:
      - events
      summary: Найти события в каталоге
      description: Применяет фильтры до пагинации и возвращает признак сохранения.
      operationId: catalog_api_v1_events_catalog_get
      security:
      - BearerAuth: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
          maxLength: 200
          description: Поиск по названию события или площадки.
          default: ''
          title: Q
        description: Поиск по названию события или площадки.
      - name: date_mode
        in: query
        required: false
        schema:
          enum:
          - any
          - today
          - weekend
          - date
          type: string
          description: Предустановленный период или конкретная дата.
          default: any
          title: Date Mode
        description: Предустановленный период или конкретная дата.
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^\d{4}-\d{2}-\d{2}$
          - type: 'null'
          description: Дата YYYY-MM-DD для режима date.
          title: Date
        description: Дата YYYY-MM-DD для режима date.
      - name: categories
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
          description: Повторяемый идентификатор категории.
          default: []
          title: Categories
        description: Повторяемый идентификатор категории.
      - name: free_only
        in: query
        required: false
        schema:
          type: boolean
          description: Показывать только бесплатные события.
          default: false
          title: Free Only
        description: Показывать только бесплатные события.
      - name: budget_max
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 1000000
            minimum: 0
          - type: 'null'
          description: Максимальная минимальная цена; несовместимо с free_only.
          title: Budget Max
        description: Максимальная минимальная цена; несовместимо с free_only.
      - name: time
        in: query
        required: false
        schema:
          enum:
          - any
          - evening
          type: string
          description: Любое время или начало не раньше 18:00.
          default: any
          title: Time
        description: Любое время или начало не раньше 18:00.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Размер страницы.
          default: 20
          title: Limit
        description: Размер страницы.
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Смещение от начала результата.
          default: 0
          title: Offset
        description: Смещение от начала результата.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Операция недоступна до завершения анкеты.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Catalog Api V1 Events Catalog Get
        '502':
          description: Внешний поставщик событий временно недоступен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
  /api/v1/events/favorites:
    get:
      tags:
      - events
      summary: Получить избранные события
      operationId: favorites_api_v1_events_favorites_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventOut'
                type: array
                title: Response Favorites Api V1 Events Favorites Get
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
  /api/v1/events/{event_id}:
    get:
      tags:
      - events
      summary: Получить событие
      operationId: detail_api_v1_events__event_id__get
      security:
      - BearerAuth: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Detail Api V1 Events  Event Id  Get
  /api/v1/events/{event_id}/reaction:
    post:
      tags:
      - events
      summary: Сохранить реакцию на событие
      operationId: react_api_v1_events__event_id__reaction_post
      security:
      - BearerAuth: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 React Api V1 Events  Event Id  Reaction Post
    delete:
      tags:
      - events
      summary: Удалить реакцию на событие
      operationId: remove_reaction_api_v1_events__event_id__reaction_delete
      security:
      - BearerAuth: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Remove Reaction Api V1 Events  Event Id  Reaction Delete
...
components:
  schemas:
...
    CatalogEventOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        category_name:
          type: string
          title: Category Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        image_credit:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Credit
        start_date:
          type: string
          format: date-time
          title: Start Date
        end_date:
          type: string
          format: date-time
          title: End Date
        timezone:
          type: string
          title: Timezone
        price_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Min
        price_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Max
        is_free:
          type: boolean
          title: Is Free
        city:
          type: string
          title: City
        location_name:
          type: string
          title: Location Name
        address:
          type: string
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        provider:
          type: string
          title: Provider
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        is_saved:
          type: boolean
          title: Is Saved
      type: object
      required:
      - id
      - title
      - description
      - category
      - category_name
      - tags
      - image_url
      - image_credit
      - start_date
      - end_date
      - timezone
      - price_min
      - price_max
      - is_free
      - city
      - location_name
      - address
      - latitude
      - longitude
      - source_url
      - provider
      - is_saved
      title: CatalogEventOut
    CatalogOut:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CatalogEventOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - has_more
      title: CatalogOut
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
      description: Ошибка бизнес-правила или доступа.
...
    EventOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        category_name:
          type: string
          title: Category Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        image_credit:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Credit
        start_date:
          type: string
          format: date-time
          title: Start Date
        end_date:
          type: string
          format: date-time
          title: End Date
        timezone:
          type: string
          title: Timezone
        price_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Min
        price_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Max
        is_free:
          type: boolean
          title: Is Free
        city:
          type: string
          title: City
        location_name:
          type: string
          title: Location Name
        address:
          type: string
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        provider:
          type: string
          title: Provider
        reasons:
          items:
            type: string
          type: array
          title: Reasons
      type: object
      required:
      - id
      - title
      - description
      - category
      - category_name
      - tags
      - image_url
      - image_credit
      - start_date
      - end_date
      - timezone
      - price_min
      - price_max
      - is_free
      - city
      - location_name
      - address
      - latitude
      - longitude
      - source_url
      - provider
      title: EventOut
...
    PreferencesInput:
      properties:
        city:
          type: string
          maxLength: 200
          minLength: 2
          title: City
        categories:
          items:
            type: string
          type: array
          maxItems: 20
          minItems: 1
          title: Categories
        budget_max:
          anyOf:
          - type: integer
            maximum: 1000000.0
            minimum: 0.0
          - type: 'null'
          title: Budget Max
        companion:
          type: string
          enum:
          - solo
          - friends
          - partner
          - family
          - children
          title: Companion
          default: friends
        preferred_days:
          type: string
          enum:
          - any
          - weekdays
          - weekends
          title: Preferred Days
          default: any
        preferred_time:
          type: string
          enum:
          - any
          - morning
          - day
          - evening
          title: Preferred Time
          default: any
      type: object
      required:
      - city
      - categories
      title: PreferencesInput
      description: Настройки персонализации и условия досуга пользователя.
      examples:
      - budget_max: 3000
        categories:
        - koncerty
        - vystavki
        city: Москва
        companion: friends
        preferred_days: weekends
        preferred_time: evening
    PreferencesOut:
      properties:
        city:
          type: string
          title: City
        categories:
          items:
            type: string
          type: array
          title: Categories
        budget_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Budget Max
        companion:
          type: string
          title: Companion
        preferred_days:
          type: string
          title: Preferred Days
        preferred_time:
          type: string
          title: Preferred Time
      type: object
      required:
      - city
      - categories
      - budget_max
      - companion
      - preferred_days
      - preferred_time
      title: PreferencesOut
    ReactionInput:
      properties:
        reaction:
          type: string
          enum:
          - like
          - dislike
          title: Reaction
      type: object
      required:
      - reaction
      title: ReactionInput
      description: Реакция пользователя на событие.
      examples:
      - reaction: like
    ReactionOut:
      properties:
        reaction:
          type: string
          title: Reaction
      type: object
      required:
      - reaction
      title: ReactionOut
      examples:
      - reaction: like
    RequestValidationErrorOut:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationIssue'
          type: array
          title: Detail
      type: object
      required:
      - detail
      title: RequestValidationErrorOut
      description: Ошибки автоматической проверки запроса FastAPI.
...
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        first_name:
          type: string
          title: First Name
        city:
          type: string
          title: City
        onboarding_completed:
          type: boolean
          title: Onboarding Completed
      type: object
      required:
      - id
      - first_name
      - city
      - onboarding_completed
      title: UserOut
    ValidationIssue:
      properties:
        type:
          type: string
          title: Type
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Loc
        msg:
          type: string
          title: Msg
        input:
          anyOf:
          - {}
          - type: 'null'
          title: Input
        ctx:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Ctx
      additionalProperties: true
      type: object
      required:
      - type
      - loc
      - msg
      title: ValidationIssue
      description: Одна ошибка проверки структуры запроса.
  securitySchemes:
    BearerAuth:
      type: http
      description: JWT, полученный через один из методов авторизации.
      scheme: bearer
    AdminToken:
      type: apiKey
      description: Секрет ручной синхронизации событий.
      in: header
      name: X-Admin-Token
...